+++ escrow/project/contracts/escrow-contract/tests/functions/core/deposit.rs
@@ -1,16 +1,24 @@
 use crate::utils::{
-    abi_calls::{create_escrow, deposit, propose_arbiter, return_deposit},
-    test_helpers::{asset_amount, create_arbiter, create_asset, mint, setup},
+    interface::core::{create_escrow, deposit},
+    setup::{create_arbiter, create_asset, mint, setup},
 };
+use fuels::tx::ContractId;
 
 mod success {
 
     use super::*;
-    use crate::utils::ReturnedDepositEvent;
+    use crate::utils::setup::{asset_amount, DepositEvent};
 
     #[tokio::test]
-    async fn returns_deposit() {
+    async fn deposits() {
         let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
 
         mint(
             seller.wallet.address(),
@@ -25,13 +33,10 @@ mod success {
         )
         .await;
 
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
+        assert_eq!(
             defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+            asset_amount(&defaults.asset_id, &buyer.wallet).await
+        );
 
         create_escrow(
             defaults.asset_amount,
@@ -43,32 +48,37 @@ mod success {
             defaults.deadline,
         )
         .await;
-        deposit(
+
+        let response = deposit(
             defaults.asset_amount,
             &defaults.asset_id,
             &buyer.contract,
             0,
         )
         .await;
-
-        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
-
-        let response = return_deposit(&seller.contract, 0).await;
-        let log = response
-            .get_logs_with_type::<ReturnedDepositEvent>()
-            .unwrap();
+        let log = response.get_logs_with_type::<DepositEvent>().unwrap();
         let event = log.get(0).unwrap();
 
-        assert_eq!(*event, ReturnedDepositEvent { identifier: 0 });
         assert_eq!(
-            defaults.asset_amount,
-            asset_amount(&defaults.asset_id, &buyer.wallet).await
+            *event,
+            DepositEvent {
+                asset: defaults.asset_id,
+                identifier: 0
+            }
         );
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
     }
 
     #[tokio::test]
-    async fn returns_deposit_after_proposing_arbiter() {
+    async fn deposits_to_two_escrows() {
         let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
 
         mint(
             seller.wallet.address(),
@@ -78,88 +88,21 @@ mod success {
         .await;
         mint(
             buyer.wallet.address(),
-            defaults.asset_amount,
+            defaults.asset_amount * 2,
             &defaults.asset,
         )
         .await;
 
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-        propose_arbiter(arbiter_obj, &seller.contract, 0).await;
-        deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-
-        assert_eq!(0, asset_amount(&defaults.asset_id, &seller.wallet).await);
-        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
-
-        let response = return_deposit(&seller.contract, 0).await;
-        let log = response
-            .get_logs_with_type::<ReturnedDepositEvent>()
-            .unwrap();
-        let event = log.get(0).unwrap();
-
-        assert_eq!(*event, ReturnedDepositEvent { identifier: 0 });
-        assert_eq!(
-            defaults.asset_amount,
-            asset_amount(&defaults.asset_id, &buyer.wallet).await
-        );
         assert_eq!(
             defaults.asset_amount * 2,
             asset_amount(&defaults.asset_id, &seller.wallet).await
         );
-    }
-
-    #[tokio::test]
-    async fn returns_deposit_in_two_escrows() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-
-        mint(
-            seller.wallet.address(),
-            defaults.asset_amount * 2,
-            &defaults.asset,
-        )
-        .await;
-        mint(
-            buyer.wallet.address(),
-            defaults.asset_amount * 2,
-            &defaults.asset,
-        )
-        .await;
-
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
 
         create_escrow(
             defaults.asset_amount,
             &arbiter_obj,
             &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
+            vec![asset.clone()],
             buyer.wallet.address(),
             &seller.contract,
             defaults.deadline,
@@ -169,59 +112,104 @@ mod success {
             defaults.asset_amount,
             &arbiter_obj,
             &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
+            vec![asset.clone(), asset.clone(), asset.clone()],
             buyer.wallet.address(),
             &seller.contract,
             defaults.deadline,
         )
         .await;
 
-        deposit(
+        let response = deposit(
             defaults.asset_amount,
             &defaults.asset_id,
             &buyer.contract,
             0,
         )
         .await;
-        deposit(
+        let log = response.get_logs_with_type::<DepositEvent>().unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(
+            *event,
+            DepositEvent {
+                asset: defaults.asset_id,
+                identifier: 0
+            }
+        );
+        assert_eq!(
+            defaults.asset_amount,
+            asset_amount(&defaults.asset_id, &buyer.wallet).await
+        );
+
+        let response = deposit(
             defaults.asset_amount,
             &defaults.asset_id,
             &buyer.contract,
             1,
         )
         .await;
+        let log = response.get_logs_with_type::<DepositEvent>().unwrap();
+        let event = log.get(0).unwrap();
 
+        assert_eq!(
+            *event,
+            DepositEvent {
+                asset: defaults.asset_id,
+                identifier: 1
+            }
+        );
         assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
-
-        let response = return_deposit(&seller.contract, 0).await;
-        let log = response
-            .get_logs_with_type::<ReturnedDepositEvent>()
-            .unwrap();
-        let event = log.get(0).unwrap();
-
-        assert_eq!(*event, ReturnedDepositEvent { identifier: 0 });
-        assert_eq!(
-            defaults.asset_amount,
-            asset_amount(&defaults.asset_id, &buyer.wallet).await
-        );
-
-        let response = return_deposit(&seller.contract, 1).await;
-        let log = response
-            .get_logs_with_type::<ReturnedDepositEvent>()
-            .unwrap();
-        let event = log.get(0).unwrap();
-
-        assert_eq!(*event, ReturnedDepositEvent { identifier: 1 });
-        assert_eq!(
-            defaults.asset_amount * 2,
-            asset_amount(&defaults.asset_id, &buyer.wallet).await
-        );
     }
 }
 
 mod revert {
 
     use super::*;
+    use crate::utils::{interface::core::transfer_to_seller, setup::create_asset_with_salt};
+
+    #[tokio::test]
+    #[should_panic(expected = "EscrowExpired")]
+    async fn when_deadline_is_reached() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            6,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
+    }
 
     #[tokio::test]
     #[should_panic(expected = "StateNotPending")]
@@ -243,7 +231,7 @@ mod revert {
         .await;
         mint(
             buyer.wallet.address(),
-            defaults.asset_amount,
+            defaults.asset_amount * 2,
             &defaults.asset,
         )
         .await;
@@ -265,13 +253,57 @@ mod revert {
             0,
         )
         .await;
-        return_deposit(&seller.contract, 0).await;
-        return_deposit(&seller.contract, 0).await;
+        transfer_to_seller(&buyer.contract, 0).await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
     }
 
     #[tokio::test]
     #[should_panic(expected = "Unauthorized")]
-    async fn when_caller_is_not_seller() {
+    async fn when_caller_is_not_buyer() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount * 2,
+            &defaults.asset,
+        )
+        .await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &seller.contract,
+            0,
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "AlreadyDeposited")]
+    async fn when_depositing_more_than_once() {
         let (arbiter, buyer, seller, defaults) = setup().await;
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -289,7 +321,7 @@ mod revert {
         .await;
         mint(
             buyer.wallet.address(),
-            defaults.asset_amount,
+            defaults.asset_amount * 2,
             &defaults.asset,
         )
         .await;
@@ -311,12 +343,18 @@ mod revert {
             0,
         )
         .await;
-        return_deposit(&buyer.contract, 0).await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
     }
 
     #[tokio::test]
-    #[should_panic(expected = "CannotTransferBeforeDesposit")]
-    async fn when_buyer_has_not_deposited() {
+    #[should_panic(expected = "IncorrectAssetAmount")]
+    async fn when_incorrect_asset_amount_is_sent() {
         let (arbiter, buyer, seller, defaults) = setup().await;
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -334,7 +372,7 @@ mod revert {
         .await;
         mint(
             buyer.wallet.address(),
-            defaults.asset_amount,
+            defaults.asset_amount * 2,
             &defaults.asset,
         )
         .await;
@@ -349,6 +387,57 @@ mod revert {
             defaults.deadline,
         )
         .await;
-        return_deposit(&seller.contract, 0).await;
+        deposit(
+            defaults.asset_amount - 1,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "IncorrectAssetSent")]
+    async fn when_incorrect_asset_is_sent() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+        let (id, salted_asset) = create_asset_with_salt([1u8; 32], buyer.wallet.clone()).await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount * 2,
+            &salted_asset,
+        )
+        .await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &ContractId::from(*id),
+            &buyer.contract,
+            0,
+        )
+        .await;
     }
 }
