+++ escrow/project/contracts/escrow-contract/tests/functions/core/return_deposit.rs
@@ -1,20 +1,28 @@
 use crate::utils::{
-    abi_calls::{create_escrow, deposit, propose_arbiter, return_deposit, withdraw_collateral},
-    test_helpers::{asset_amount, create_arbiter, create_asset, mint, setup},
+    interface::core::{create_escrow, deposit, return_deposit},
+    setup::{create_arbiter, create_asset, mint, setup},
 };
 
 mod success {
 
     use super::*;
-    use crate::utils::WithdrawnCollateralEvent;
+    use crate::utils::{
+        interface::core::propose_arbiter,
+        setup::{asset_amount, ReturnedDepositEvent},
+    };
 
     #[tokio::test]
-    #[ignore]
-    async fn withdraws_collateral() {
+    async fn returns_deposit() {
         let (arbiter, buyer, seller, defaults) = setup().await;
 
         mint(
             seller.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
             defaults.asset_amount,
             &defaults.asset,
         )
@@ -35,28 +43,34 @@ mod success {
             vec![asset.clone(), asset.clone()],
             buyer.wallet.address(),
             &seller.contract,
-            6,
+            defaults.deadline,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
         )
         .await;
 
-        assert_eq!(0, asset_amount(&defaults.asset_id, &seller.wallet).await);
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
 
-        // TODO: need to shift block by one, waiting on SDK
-        let response = withdraw_collateral(&seller.contract, 0).await;
+        let response = return_deposit(&seller.contract, 0).await;
         let log = response
-            .get_logs_with_type::<WithdrawnCollateralEvent>()
+            .get_logs_with_type::<ReturnedDepositEvent>()
             .unwrap();
         let event = log.get(0).unwrap();
 
-        assert_eq!(*event, WithdrawnCollateralEvent { identifier: 0 });
+        assert_eq!(*event, ReturnedDepositEvent { identifier: 0 });
         assert_eq!(
             defaults.asset_amount,
-            asset_amount(&defaults.asset_id, &seller.wallet).await
+            asset_amount(&defaults.asset_id, &buyer.wallet).await
         );
     }
 
     #[tokio::test]
-    async fn withdraws_collateral_after_proposing_arbiter() {
+    async fn returns_deposit_after_proposing_arbiter() {
         let (arbiter, buyer, seller, defaults) = setup().await;
 
         mint(
@@ -65,6 +79,12 @@ mod success {
             &defaults.asset,
         )
         .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
 
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -81,21 +101,32 @@ mod success {
             vec![asset.clone(), asset.clone()],
             buyer.wallet.address(),
             &seller.contract,
-            5,
+            defaults.deadline,
         )
         .await;
-
         propose_arbiter(arbiter_obj, &seller.contract, 0).await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
 
         assert_eq!(0, asset_amount(&defaults.asset_id, &seller.wallet).await);
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
 
-        let response = withdraw_collateral(&seller.contract, 0).await;
+        let response = return_deposit(&seller.contract, 0).await;
         let log = response
-            .get_logs_with_type::<WithdrawnCollateralEvent>()
+            .get_logs_with_type::<ReturnedDepositEvent>()
             .unwrap();
         let event = log.get(0).unwrap();
 
-        assert_eq!(*event, WithdrawnCollateralEvent { identifier: 0 });
+        assert_eq!(*event, ReturnedDepositEvent { identifier: 0 });
+        assert_eq!(
+            defaults.asset_amount,
+            asset_amount(&defaults.asset_id, &buyer.wallet).await
+        );
         assert_eq!(
             defaults.asset_amount * 2,
             asset_amount(&defaults.asset_id, &seller.wallet).await
@@ -103,9 +134,91 @@ mod success {
     }
 
     #[tokio::test]
-    #[ignore]
-    async fn withdraws_collateral_in_two_escrows() {
-        // TODO: skipping similar to withdraws_collateral
+    async fn returns_deposit_in_two_escrows() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount * 2,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount * 2,
+            &defaults.asset,
+        )
+        .await;
+
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            1,
+        )
+        .await;
+
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+
+        let response = return_deposit(&seller.contract, 0).await;
+        let log = response
+            .get_logs_with_type::<ReturnedDepositEvent>()
+            .unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(*event, ReturnedDepositEvent { identifier: 0 });
+        assert_eq!(
+            defaults.asset_amount,
+            asset_amount(&defaults.asset_id, &buyer.wallet).await
+        );
+
+        let response = return_deposit(&seller.contract, 1).await;
+        let log = response
+            .get_logs_with_type::<ReturnedDepositEvent>()
+            .unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(*event, ReturnedDepositEvent { identifier: 1 });
+        assert_eq!(
+            defaults.asset_amount * 2,
+            asset_amount(&defaults.asset_id, &buyer.wallet).await
+        );
     }
 }
 
@@ -156,59 +269,12 @@ mod revert {
         )
         .await;
         return_deposit(&seller.contract, 0).await;
-        withdraw_collateral(&seller.contract, 0).await;
+        return_deposit(&seller.contract, 0).await;
     }
 
     #[tokio::test]
-    #[should_panic(expected = "CannotWithdrawBeforeDeadline")]
-    async fn when_deadline_is_not_in_the_past() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-
-        mint(
-            seller.wallet.address(),
-            defaults.asset_amount,
-            &defaults.asset,
-        )
-        .await;
-        mint(
-            buyer.wallet.address(),
-            defaults.asset_amount,
-            &defaults.asset,
-        )
-        .await;
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-        deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-        withdraw_collateral(&seller.contract, 0).await;
-    }
-
-    #[tokio::test]
-    #[ignore]
     #[should_panic(expected = "Unauthorized")]
     async fn when_caller_is_not_seller() {
-        // Test passes when deadline requirement is met. Ignored till SDK manipulation to prevent failure
         let (arbiter, buyer, seller, defaults) = setup().await;
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -248,14 +314,12 @@ mod revert {
             0,
         )
         .await;
-        withdraw_collateral(&seller.contract, 0).await;
+        return_deposit(&buyer.contract, 0).await;
     }
 
     #[tokio::test]
-    #[ignore]
-    #[should_panic(expected = "CannotWithdrawAfterDesposit")]
-    async fn when_buyer_has_deposited() {
-        // Test passes when deadline requirement is met. Ignored till SDK manipulation to prevent failure
+    #[should_panic(expected = "CannotTransferBeforeDesposit")]
+    async fn when_buyer_has_not_deposited() {
         let (arbiter, buyer, seller, defaults) = setup().await;
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -288,13 +352,6 @@ mod revert {
             defaults.deadline,
         )
         .await;
-        deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-        withdraw_collateral(&seller.contract, 0).await;
+        return_deposit(&seller.contract, 0).await;
     }
 }
