+++ escrow/project/contracts/escrow-contract/tests/functions/core/transfer_to_seller.rs
@@ -1,176 +1,18 @@
 use crate::utils::{
-    abi_calls::{create_escrow, deposit, transfer_to_seller},
-    test_helpers::{
-        asset_amount, create_arbiter, create_asset, create_asset_with_salt, mint, setup,
-    },
+    interface::core::{create_escrow, deposit, transfer_to_seller},
+    setup::{create_arbiter, create_asset, mint, setup},
 };
-use fuels::tx::ContractId;
 
 mod success {
 
     use super::*;
-    use crate::utils::DepositEvent;
+    use crate::utils::{
+        interface::core::propose_arbiter,
+        setup::{asset_amount, TransferredToSellerEvent},
+    };
 
     #[tokio::test]
-    async fn deposits() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-
-        mint(
-            seller.wallet.address(),
-            defaults.asset_amount,
-            &defaults.asset,
-        )
-        .await;
-        mint(
-            buyer.wallet.address(),
-            defaults.asset_amount,
-            &defaults.asset,
-        )
-        .await;
-
-        assert_eq!(
-            defaults.asset_amount,
-            asset_amount(&defaults.asset_id, &buyer.wallet).await
-        );
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-
-        let response = deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-        let log = response.get_logs_with_type::<DepositEvent>().unwrap();
-        let event = log.get(0).unwrap();
-
-        assert_eq!(
-            *event,
-            DepositEvent {
-                asset: defaults.asset_id,
-                identifier: 0
-            }
-        );
-        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
-    }
-
-    #[tokio::test]
-    async fn deposits_to_two_escrows() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-
-        mint(
-            seller.wallet.address(),
-            defaults.asset_amount * 2,
-            &defaults.asset,
-        )
-        .await;
-        mint(
-            buyer.wallet.address(),
-            defaults.asset_amount * 2,
-            &defaults.asset,
-        )
-        .await;
-
-        assert_eq!(
-            defaults.asset_amount * 2,
-            asset_amount(&defaults.asset_id, &seller.wallet).await
-        );
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-
-        let response = deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-        let log = response.get_logs_with_type::<DepositEvent>().unwrap();
-        let event = log.get(0).unwrap();
-
-        assert_eq!(
-            *event,
-            DepositEvent {
-                asset: defaults.asset_id,
-                identifier: 0
-            }
-        );
-        assert_eq!(
-            defaults.asset_amount,
-            asset_amount(&defaults.asset_id, &buyer.wallet).await
-        );
-
-        let response = deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            1,
-        )
-        .await;
-        let log = response.get_logs_with_type::<DepositEvent>().unwrap();
-        let event = log.get(0).unwrap();
-
-        assert_eq!(
-            *event,
-            DepositEvent {
-                asset: defaults.asset_id,
-                identifier: 1
-            }
-        );
-        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
-    }
-}
-
-mod revert {
-
-    use super::*;
-
-    #[tokio::test]
-    #[should_panic(expected = "EscrowExpired")]
-    async fn when_deadline_is_reached() {
+    async fn transfers_to_seller() {
         let (arbiter, buyer, seller, defaults) = setup().await;
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -200,7 +42,7 @@ mod revert {
             vec![asset.clone(), asset.clone()],
             buyer.wallet.address(),
             &seller.contract,
-            6,
+            defaults.deadline,
         )
         .await;
         deposit(
@@ -210,7 +52,178 @@ mod revert {
             0,
         )
         .await;
+
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+        assert_eq!(0, asset_amount(&defaults.asset_id, &seller.wallet).await);
+
+        let response = transfer_to_seller(&buyer.contract, 0).await;
+        let log = response
+            .get_logs_with_type::<TransferredToSellerEvent>()
+            .unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(*event, TransferredToSellerEvent { identifier: 0 });
+        assert_eq!(
+            defaults.asset_amount * 2,
+            asset_amount(&defaults.asset_id, &seller.wallet).await
+        );
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
     }
+
+    #[tokio::test]
+    async fn transfers_to_seller_after_proposing_arbiter() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount * 2,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
+        propose_arbiter(arbiter_obj, &seller.contract, 0).await;
+
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+        assert_eq!(0, asset_amount(&defaults.asset_id, &seller.wallet).await);
+
+        let response = transfer_to_seller(&buyer.contract, 0).await;
+        let log = response
+            .get_logs_with_type::<TransferredToSellerEvent>()
+            .unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(*event, TransferredToSellerEvent { identifier: 0 });
+        assert_eq!(
+            defaults.asset_amount * 3,
+            asset_amount(&defaults.asset_id, &seller.wallet).await
+        );
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+    }
+
+    #[tokio::test]
+    async fn transfers_to_seller_in_two_escrows() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
+            defaults.asset_amount,
+        )
+        .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount * 2,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount * 2,
+            &defaults.asset,
+        )
+        .await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            0,
+        )
+        .await;
+        deposit(
+            defaults.asset_amount,
+            &defaults.asset_id,
+            &buyer.contract,
+            1,
+        )
+        .await;
+
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+        assert_eq!(0, asset_amount(&defaults.asset_id, &seller.wallet).await);
+
+        let response = transfer_to_seller(&buyer.contract, 0).await;
+        let log = response
+            .get_logs_with_type::<TransferredToSellerEvent>()
+            .unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(*event, TransferredToSellerEvent { identifier: 0 });
+        assert_eq!(
+            defaults.asset_amount * 2,
+            asset_amount(&defaults.asset_id, &seller.wallet).await
+        );
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+
+        let response = transfer_to_seller(&buyer.contract, 1).await;
+        let log = response
+            .get_logs_with_type::<TransferredToSellerEvent>()
+            .unwrap();
+        let event = log.get(0).unwrap();
+
+        assert_eq!(*event, TransferredToSellerEvent { identifier: 1 });
+        assert_eq!(
+            defaults.asset_amount * 4,
+            asset_amount(&defaults.asset_id, &seller.wallet).await
+        );
+        assert_eq!(0, asset_amount(&defaults.asset_id, &buyer.wallet).await);
+    }
+}
+
+mod revert {
+
+    use super::*;
 
     #[tokio::test]
     #[should_panic(expected = "StateNotPending")]
@@ -232,7 +245,7 @@ mod revert {
         .await;
         mint(
             buyer.wallet.address(),
-            defaults.asset_amount * 2,
+            defaults.asset_amount,
             &defaults.asset,
         )
         .await;
@@ -255,13 +268,45 @@ mod revert {
         )
         .await;
         transfer_to_seller(&buyer.contract, 0).await;
-        deposit(
+        transfer_to_seller(&buyer.contract, 0).await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "CannotTransferBeforeDesposit")]
+    async fn when_buyer_has_not_deposited() {
+        let (arbiter, buyer, seller, defaults) = setup().await;
+        let arbiter_obj = create_arbiter(
+            arbiter.wallet.address(),
+            defaults.asset_id,
             defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
         )
         .await;
+        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
+
+        mint(
+            seller.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+        mint(
+            buyer.wallet.address(),
+            defaults.asset_amount,
+            &defaults.asset,
+        )
+        .await;
+
+        create_escrow(
+            defaults.asset_amount,
+            &arbiter_obj,
+            &defaults.asset_id,
+            vec![asset.clone(), asset.clone()],
+            buyer.wallet.address(),
+            &seller.contract,
+            defaults.deadline,
+        )
+        .await;
+        transfer_to_seller(&buyer.contract, 0).await;
     }
 
     #[tokio::test]
@@ -278,51 +323,13 @@ mod revert {
 
         mint(
             seller.wallet.address(),
-            defaults.asset_amount * 2,
-            &defaults.asset,
-        )
-        .await;
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-        deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &seller.contract,
-            0,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "AlreadyDeposited")]
-    async fn when_depositing_more_than_once() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-
-        mint(
-            seller.wallet.address(),
             defaults.asset_amount,
             &defaults.asset,
         )
         .await;
         mint(
             buyer.wallet.address(),
-            defaults.asset_amount * 2,
+            defaults.asset_amount,
             &defaults.asset,
         )
         .await;
@@ -344,101 +351,6 @@ mod revert {
             0,
         )
         .await;
-        deposit(
-            defaults.asset_amount,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "IncorrectAssetAmount")]
-    async fn when_incorrect_asset_amount_is_sent() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-
-        mint(
-            seller.wallet.address(),
-            defaults.asset_amount,
-            &defaults.asset,
-        )
-        .await;
-        mint(
-            buyer.wallet.address(),
-            defaults.asset_amount * 2,
-            &defaults.asset,
-        )
-        .await;
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-        deposit(
-            defaults.asset_amount - 1,
-            &defaults.asset_id,
-            &buyer.contract,
-            0,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "IncorrectAssetSent")]
-    async fn when_incorrect_asset_is_sent() {
-        let (arbiter, buyer, seller, defaults) = setup().await;
-        let arbiter_obj = create_arbiter(
-            arbiter.wallet.address(),
-            defaults.asset_id,
-            defaults.asset_amount,
-        )
-        .await;
-        let asset = create_asset(defaults.asset_amount, defaults.asset_id).await;
-        let (id, salted_asset) = create_asset_with_salt([1u8; 32], buyer.wallet.clone()).await;
-
-        mint(
-            seller.wallet.address(),
-            defaults.asset_amount,
-            &defaults.asset,
-        )
-        .await;
-        mint(
-            buyer.wallet.address(),
-            defaults.asset_amount * 2,
-            &salted_asset,
-        )
-        .await;
-
-        create_escrow(
-            defaults.asset_amount,
-            &arbiter_obj,
-            &defaults.asset_id,
-            vec![asset.clone(), asset.clone()],
-            buyer.wallet.address(),
-            &seller.contract,
-            defaults.deadline,
-        )
-        .await;
-        deposit(
-            defaults.asset_amount,
-            &ContractId::from(*id),
-            &buyer.contract,
-            0,
-        )
-        .await;
+        transfer_to_seller(&seller.contract, 0).await;
     }
 }
