+++ escrow/project/contracts/escrow-contract/tests/functions/mod.rs
@@ -1,10 +1,2 @@
-mod accept_arbiter;
-mod create_escrow;
-mod deposit;
-mod dispute;
-mod propose_arbiter;
-mod resolve_dispute;
-mod return_deposit;
-mod take_payment;
-mod transfer_to_seller;
-mod withdraw_collateral;
+mod core;
+mod info;
